+++ Sources/NablaMessagingCore/Domain/Interactors/CreateConversation/CreateConversationInteractor.swift
@@ -1,5 +1,9 @@
 import Foundation
 
 protocol CreateConversationInteractor {
-    func execute(handler: ResultHandler<Conversation, NablaError>) -> Cancellable
+    func execute(
+        title: String?,
+        providerIdToAssign: UUID?,
+        handler: ResultHandler<Conversation, NablaError>
+    ) -> Cancellable
 }