+++ Sources/NablaMessagingCore/Domain/Repositories/ConversationRepository.swift
@@ -3,5 +3,9 @@ import Foundation
 protocol ConversationRepository {
     func watchConversation(_ conversationId: UUID, handler: ResultHandler<Conversation, NablaError>) -> Cancellable
     func watchConversations(handler: ResultHandler<ConversationList, NablaError>) -> PaginatedWatcher
-    func createConversation(handler: ResultHandler<Conversation, NablaError>) -> Cancellable
+    func createConversation(
+        title: String?,
+        providerIdToAssign: UUID?,
+        handler: ResultHandler<Conversation, NablaError>
+    ) -> Cancellable
 }