+++ Sources/NablaMessagingCore/NablaMessagingClient.swift
@@ -11,10 +11,21 @@ public class NablaMessagingClient {
     }
 
     /// Create a new conversation on behalf of the current user.
+    /// - Parameter title: optional - title for the conversation
+    /// - Parameter providerIdToAssign: optional - id of a provider to which the conversation should be assigned. Make sure the specified provider has enough rights to be assigned a conversation. See [Roles and Permissions](https://docs.nabla.com/docs/roles-and-permissions).
     /// - Parameter handler: Handler called when the ``Conversation`` is created or an ``Error`` if something went wrong.
     /// - Returns: A ``Cancellable`` of the task
-    public func createConversation(handler: @escaping (Result<Conversation, NablaError>) -> Void) -> Cancellable {
-        container.createConversationInteractor.execute(handler: .init(handler))
+    public func createConversation(
+        title: String?,
+        providerIdToAssign: UUID?,
+        handler: @escaping (Result<Conversation, NablaError>
+        ) -> Void
+    ) -> Cancellable {
+        container.createConversationInteractor.execute(
+            title: title,
+            providerIdToAssign: providerIdToAssign,
+            handler: .init(handler)
+        )
     }
 
     /// Watch the list of messages in a conversation.